Fix a bunch of messy stuff
