+++ src/components/pages/Dashboard.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { useHistory } from "react-router";
-import { auth, db } from "../../util/firebase";
-import { getFirestore, collection, getDocs, doc, addDoc, query, where } from "firebase/firestore";
+import { auth } from "../../util/firebaseAuth";
+import { getUserProfile } from '../../util/firebaseFirestore';
 
 function Dashboard() {
   const [user, loading] = useAuthState(auth);
@@ -17,13 +17,8 @@ function Dashboard() {
 
   const fetchUserName = async () => {
     try {
-      // const query = await db
-      //   .collection("users")
-      //   .where("uid", "==", user?.uid)
-      //   .get();
-      const q = query(collection(db, 'users'), where('uid', '==', user?.uid))
-      const data = await getDocs(q)
-      setName(data.name);
+      const profile = await getUserProfile(user)
+      setName(profile.name)
     } catch (err) {
       console.error(err);
       alert("An error occured while fetching user data");
